Replaced source C utilities with python ones
elf2binsizeword: Takes an elf, converts it to binary and replaces the
first word with the file size. Used for bootloader payloads.

elf2vlogarray: Takes and elf, converts it to binary and outputs it in
Verilog case format. Used for wb_rom.

bin2memh: Takes a binary file and output a memh readable file. Used
for the Verilator testbench.