Updated diag program for newlines
We need to print \r\n for newlines. Added some functions to
take care of that - not supernice but I think this allows for
best flexibility for the future.

Updated program.tcl to do the right things. Apparently programming
the MCS "fails" even when it succeeds so the last step is not executed.
Needs investigation.

Currently the design boots and runs the diagnostic just fine,
but due to the RAM address cycle the write test will kill the diag
program.

But it works! WOOO!